auth/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
passlib[bcrypt]==1.7.4
pydantic==2.8.2
python-jose==3.3.0
